backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
anthropic==0.21.3
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
pandas==2.2.1
scipy==1.12.0
PyYAML==6.0.1
aiofiles==23.2.1
